refactor: invert Nostr gateway dependency (DIP)
`NostrCommand`, `CommandError`, and the subscription `Message` were
defined in `infrastructure::subscription::nostr` but consumed by
`application` (state, message) and `model::nostr`, creating upward
`application -> infrastructure` and `model -> infrastructure`
dependencies (cycles, since infrastructure depends back on model/domain).

Move these contract types into a new `model::nostr_gateway` module and
have the infrastructure adapter depend inward on them. The adapter
(`NostrEvents`) stays in infrastructure.

The contract lives in `model` rather than `domain` on purpose: it
references `TimelineTabType`, a UI/model concept that the domain layer is
deliberately kept free of (see the note in
`domain::nostr::timeline_filter`). `model::nostr` is already the gateway
client (it holds the command sender), so the protocol types belong
alongside it.

Result: `application -> infrastructure` and `model -> infrastructure` are
gone; infrastructure now depends downward on `model::nostr_gateway`.
