+++ src/model.rs
@@ -1,5 +1,6 @@
 pub mod editor;
 pub mod fps;
 pub mod nostr;
+pub mod nostr_gateway;
 pub mod status_bar;
 pub mod timeline;
